refactor(arika): move EarthPoleBridge/EarthFrameBridge from orts to arika
These traits expressed pure coordinate facts about a frame — Earth's rotation
pole direction (CIP for Gcrs, +Z for SimpleEci) and the paired ECI<->ECEF
rotation / geodetic conversion — but lived in orts/src/environment.rs as a thin
dispatch over arika primitives (cip_xy, Rotation, Geodetic, EOP traits). They
belong in arika, which owns frames, rotations, and the EOP traits; orts force
models just consume them.

- arika::earth::frame_bridge: EarthPoleBridge, EarthFrameBridge + SimpleEci/Gcrs
  impls, re-exported as arika::earth::{EarthPoleBridge, EarthFrameBridge}.
- arika::earth::eop: PositionEop (position-only EOP bound, sibling of
  FullEopProvider) and GcrsEopStorage (boxed provider). Send+Sync now sits on
  the boxed storage rather than on PositionEop, matching FullEopProvider.
- The Gcrs EarthFrameBridge impl + GcrsEopStorage are #[cfg(feature = "alloc")]
  (boxed), so arika's no_std-no_alloc tier still builds; earth_pole uses
  crate::math::F64Ext + a manual clamp instead of std-only f64::max/sqrt.
- orts: repoint imports to arika::earth; delete orts/src/environment.rs.
- Regenerate the NullEop trybuild .stderr snapshots: GcrsEopStorage now appears
  in the "other types implement <EOP trait>" diagnostic note.

Behavior-preserving (pure move). Validated: arika builds std / no_std+alloc /
no_std / wasm32 with no warnings; arika + orts tests pass; clippy/fmt clean.
