+++ arika/tests/trybuild/null_eop_in_iau2006.stderr
@@ -6,11 +6,16 @@ error[E0277]: the trait bound `NullEop: NutationCorrections` is not satisfied
    |              |
    |              required by a bound introduced by this call
    |
-help: the trait `NutationCorrections` is implemented for `EopTable`
-  --> src/earth/eop/table.rs
+help: the following other types implement trait `NutationCorrections`
+  --> src/earth/eop/mod.rs
+   |
+   | impl NutationCorrections for GcrsEopStorage {
+   | ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ `GcrsEopStorage`
+   |
+  ::: src/earth/eop/table.rs
    |
    | impl NutationCorrections for EopTable {
-   | ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+   | ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ `EopTable`
 note: required by a bound in `cio_chain::<impl Rotation<arika::frame::Gcrs, arika::frame::Cirs>>::iau2006`
   --> src/earth/iau2006/cio_chain.rs
    |
