+++ arika/tests/trybuild/null_eop_in_iau2006_full_from_utc.stderr
@@ -6,11 +6,16 @@ error[E0277]: the trait bound `NullEop: Ut1Offset` is not satisfied
    |              |
    |              required by a bound introduced by this call
    |
-help: the trait `Ut1Offset` is implemented for `EopTable`
-  --> src/earth/eop/table.rs
+help: the following other types implement trait `Ut1Offset`
+  --> src/earth/eop/mod.rs
+   |
+   | impl Ut1Offset for GcrsEopStorage {
+   | ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ `GcrsEopStorage`
+   |
+  ::: src/earth/eop/table.rs
    |
    | impl Ut1Offset for EopTable {
-   | ^^^^^^^^^^^^^^^^^^^^^^^^^^^
+   | ^^^^^^^^^^^^^^^^^^^^^^^^^^^ `EopTable`
 note: required by a bound in `cio_chain::<impl Rotation<arika::frame::Gcrs, arika::frame::Itrs>>::iau2006_full_from_utc`
   --> src/earth/iau2006/cio_chain.rs
    |
@@ -28,11 +33,16 @@ error[E0277]: the trait bound `NullEop: NutationCorrections` is not satisfied
    |              |
    |              required by a bound introduced by this call
    |
-help: the trait `NutationCorrections` is implemented for `EopTable`
-  --> src/earth/eop/table.rs
+help: the following other types implement trait `NutationCorrections`
+  --> src/earth/eop/mod.rs
+   |
+   | impl NutationCorrections for GcrsEopStorage {
+   | ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ `GcrsEopStorage`
+   |
+  ::: src/earth/eop/table.rs
    |
    | impl NutationCorrections for EopTable {
-   | ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+   | ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ `EopTable`
 note: required by a bound in `cio_chain::<impl Rotation<arika::frame::Gcrs, arika::frame::Itrs>>::iau2006_full_from_utc`
   --> src/earth/iau2006/cio_chain.rs
    |
@@ -50,11 +60,16 @@ error[E0277]: the trait bound `NullEop: PolarMotion` is not satisfied
    |              |
    |              required by a bound introduced by this call
    |
-help: the trait `PolarMotion` is implemented for `EopTable`
-  --> src/earth/eop/table.rs
+help: the following other types implement trait `PolarMotion`
+  --> src/earth/eop/mod.rs
+   |
+   | impl PolarMotion for GcrsEopStorage {
+   | ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ `GcrsEopStorage`
+   |
+  ::: src/earth/eop/table.rs
    |
    | impl PolarMotion for EopTable {
-   | ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+   | ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ `EopTable`
 note: required by a bound in `cio_chain::<impl Rotation<arika::frame::Gcrs, arika::frame::Itrs>>::iau2006_full_from_utc`
   --> src/earth/iau2006/cio_chain.rs
    |
