+++ arika/tests/trybuild/null_eop_in_polar_motion.stderr
@@ -6,11 +6,16 @@ error[E0277]: the trait bound `NullEop: PolarMotion` is not satisfied
    |              |
    |              required by a bound introduced by this call
    |
-help: the trait `PolarMotion` is implemented for `EopTable`
-  --> src/earth/eop/table.rs
+help: the following other types implement trait `PolarMotion`
+  --> src/earth/eop/mod.rs
+   |
+   | impl PolarMotion for GcrsEopStorage {
+   | ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ `GcrsEopStorage`
+   |
+  ::: src/earth/eop/table.rs
    |
    | impl PolarMotion for EopTable {
-   | ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+   | ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ `EopTable`
 note: required by a bound in `cio_chain::<impl Rotation<arika::frame::Tirs, arika::frame::Itrs>>::polar_motion`
   --> src/earth/iau2006/cio_chain.rs
    |
