+++ arika/tests/trybuild/null_eop_in_to_ut1.stderr
@@ -6,11 +6,16 @@ error[E0277]: the trait bound `NullEop: Ut1Offset` is not satisfied
    |                    |
    |                    required by a bound introduced by this call
    |
-help: the trait `Ut1Offset` is implemented for `EopTable`
-  --> src/earth/eop/table.rs
+help: the following other types implement trait `Ut1Offset`
+  --> src/earth/eop/mod.rs
+   |
+   | impl Ut1Offset for GcrsEopStorage {
+   | ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ `GcrsEopStorage`
+   |
+  ::: src/earth/eop/table.rs
    |
    | impl Ut1Offset for EopTable {
-   | ^^^^^^^^^^^^^^^^^^^^^^^^^^^
+   | ^^^^^^^^^^^^^^^^^^^^^^^^^^^ `EopTable`
 note: required by a bound in `epoch::convert::<impl Epoch<Utc, P>>::to_ut1`
   --> src/epoch/convert.rs
    |
