+++ orts/tests/oracle_gcrf.rs
@@ -10,13 +10,13 @@
 //! - **Known EOP gap**: ZeroEop vs real EOP introduces ~arcsec-level rotation
 //!   difference (~meter-level position error per orbit)
 
+use arika::earth::GcrsEopStorage;
 use arika::earth::eop::{NutationCorrections, PolarMotion, Ut1Offset};
 use arika::earth::{J2 as J2_EARTH, J3 as J3_EARTH, J4 as J4_EARTH};
 use arika::earth::{MU as MU_EARTH, R as R_EARTH};
 use arika::epoch::Epoch;
 use arika::frame;
 use nalgebra::Vector3;
-use orts::environment::GcrsEopStorage;
 use orts::orbital::OrbitalState;
 use orts::orbital::OrbitalSystem;
 use orts::orbital::gravity::PointMass;
